refactor move createSurvey to lib
