+++ src/bot.js
@@ -1,4 +1,5 @@
 import surveyToText from "./lib/surveyToText.js";
+import { createSurvey } from "./lib/createSurvey.js";
 import {
   maxResponsesForMultiResponsePerUser,
   create_multi_cmd,
@@ -505,28 +506,4 @@ const runTest = async (redisClient) => {
   );
   await respond(redisClient, "test-survey", "evan", "comment1");
   await respond(redisClient, "test-survey", "bob", "comment2");
-};
-
-const createSurvey = async (
-  redisClient,
-  surveyName,
-  surveyType,
-  description,
-  username,
-) => {
-  await redisClient.sAdd("surveys", surveyName);
-  const initialSummaryJSON = JSON.stringify({});
-  await redisClient.set(`survey:${surveyName}:summary`, initialSummaryJSON);
-  await redisClient.set(`survey:${surveyName}:type`, surveyType);
-  await redisClient.set(`survey:${surveyName}:title`, surveyName);
-  await redisClient.set(`survey:${surveyName}:description`, description);
-  await redisClient.set(`survey:${surveyName}:username`, username);
-  await redisClient.set(`survey:${surveyName}:last-edit-time`, Date.now());
-  await redisClient.set(`survey:${surveyName}:last-summary-time`, Date.now());
-};
-
-const respond = async (redisClient, surveyName, username, response) => {
-  await redisClient.hSet(`survey:${surveyName}:responses`, username, response);
-  await redisClient.set(`survey:${surveyName}:last-edit-time`, Date.now());
-  await redisClient.publish("survey-refresh", surveyName);
 };
